requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
python-dotenv==1.0.1
Werkzeug==3.0.3
openai==1.55.3
cryptography==46.0.3
